python/pyproject.toml
[tool.poetry]
name = "ramp"
version = "0.1.0"
# TODO Bring all of this from setup.py
description = ""
# TODO Many others
authors = ["Dustin Carlino <dcarlino@turing.ac.uk>", "hsalat@turing.ac.uk", "fbenitez@turing.ac.uk"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "=1.0.3"
click = "^8.0.3"
PyYAML = "^6.0"
requests = "^2.26.0"
geopandas = "=0.7.0"
tqdm = "^4.62.3"
sklearn = "^0.0"
matplotlib = "=3.1.3"
glfw = "^2.4.0"
imgui = "^1.4.1"
PyOpenGL = "=3.1.5"
pyopencl = "=2020.2.2"
convertbng = "=0.6.25"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
